fix: correções nos metodos update e delete. Detalhes: Ambos os metodos não apresentavam a verificação caso um aluno não existisse ou seja havia uma possibilidade dele vir como null  na busca. O metodo update ainda acredito que tenha como melhorar lançando uma exception ou algo especifico, Mas para fins de teste acredito que por enquanto a como ser usado

+++ src/main/java/com/example/aula001/service/AlunoService.java
@@ -41,12 +41,24 @@ public class AlunoService {
         AlunoDTO alunoDTO = getById(id);
         Aluno aluno = new Aluno(alunoDTO);
         Optional<AlunoDTO> existingAluno = Optional.ofNullable(getById(id));
-        alunoRepository.delete(aluno);
+        if(existingAluno.isPresent()){
+            alunoRepository.delete(aluno);
+        }else {
+            throw new RuntimeException();
+        }
+
 
     }
 
     public AlunoDTO update(Aluno aluno){
         AlunoDTO alunoDTO = getById(aluno.getId());
-        return alunoDTO;
+        Optional<AlunoDTO> existingAluno = Optional.ofNullable(getById(aluno.getId()));
+        if(existingAluno.isPresent()){
+            alunoRepository.save(aluno);
+            return alunoDTO;
+        }else{
+            return null;
+        }
+
     }
 }
